1 add vad 2 silent some warnings

+++ microphone/microphone.hpp
@@ -18,7 +18,7 @@ public:
   static std::string toString(const WAVEINCAPS &x);
 
   // open microphone
-  Microphone(size_t no, const wav::Format &format, size_t voicelength,
+  Microphone(uint32_t no, const wav::Format &format, size_t voicelength,
              std::function<void(std::string &&voice)> callback) noexcept(false);
 
   // close microphone
@@ -30,27 +30,27 @@ private:
   static void CALLBACK callbackFunc(HWAVEIN hwavein, UINT uMsg,
                                     DWORD_PTR dwInstance, DWORD_PTR dwParam1,
                                     DWORD_PTR dwParam2);
-  void prepareBufferProc(size_t no);
+  void prepareBufferProc(uint32_t no);
 
 private:
   bool buffer_ready_ = false;
   bool stop_ = false;
   std::thread prepare_buffer_thread_;
-  WAVEFORMATEX format_ = { 0 };
+  WAVEFORMATEX format_ = {0};
   size_t voicelength_ = 0;
   std::function<void(std::string &&voice)> callback_;
   std::mutex mtx_;
   std::condition_variable cond_;
   HWAVEIN wave_in_ = 0;
-  std::deque<std::unique_ptr<WAVEHDR> > buffers_;
+  std::deque<std::unique_ptr<WAVEHDR>> buffers_;
 };
 
-inline void Microphone::prepareBufferProc(size_t no) {
+inline void Microphone::prepareBufferProc(uint32_t no) {
   try {
 
     MYSPACE_DEFER(stop_ = true; cond_.notify_all());
     auto n =
-        ::waveInOpen(&wave_in_, no, &format_,
+        ::waveInOpen(&wave_in_, (uint32_t)no, &format_,
                      reinterpret_cast<DWORD_PTR>(&Microphone::callbackFunc),
                      (DWORD_PTR) nullptr, CALLBACK_FUNCTION);
     MYSPACE_THROW_IF_EX(MicError, MMSYSERR_NOERROR != n, " n = ", n);
@@ -59,15 +59,15 @@ inline void Microphone::prepareBufferProc(size_t no) {
         std::max((size_t)3, (size_t)format_.nAvgBytesPerSec / voicelength_);
 
     for (size_t i = 0; i < buff_count; ++i) {
-      auto buff = std::make_unique<WAVEHDR>(WAVEHDR{ 0 });
+      auto buff = std::make_unique<WAVEHDR>(WAVEHDR{0});
 
       buff->lpData = (LPSTR)::malloc(voicelength_);
       MYSPACE_THROW_IF_EX(MicError, !buff->lpData);
-      buff->dwBufferLength = voicelength_;
+      buff->dwBufferLength = (DWORD)voicelength_;
       buff->dwBytesRecorded = 0;
       buff->dwFlags = 0;
       buff->dwLoops = 0;
-      buff->dwUser = (DWORD_PTR) this;
+      buff->dwUser = (DWORD_PTR)this;
       buff->lpNext = 0;
 
       MYSPACE_THROW_IF_EX(
@@ -99,8 +99,7 @@ inline void Microphone::prepareBufferProc(size_t no) {
         if (buffer->dwFlags & WHDR_DONE) {
           try {
             callback_(std::string(buffer->lpData, buffer->dwBytesRecorded));
-          }
-          catch (...) {
+          } catch (...) {
             MYSPACE_DEV_EXCEPTION();
           }
 
@@ -116,8 +115,7 @@ inline void Microphone::prepareBufferProc(size_t no) {
         }
       }
     }
-  }
-  catch (...) {
+  } catch (...) {
     MYSPACE_DEV_EXCEPTION();
   }
 
@@ -138,7 +136,7 @@ inline WAVEINCAPS Microphone::getMicInfo(size_t no) noexcept(false) {
   return cap;
 }
 inline Microphone::Microphone(
-    size_t no, const wav::Format &t_format, size_t voicelength,
+    uint32_t no, const wav::Format &t_format, size_t voicelength,
     std::function<void(std::string &&voice)> callback) noexcept(false)
     : voicelength_(voicelength), callback_(callback) {
 
@@ -154,9 +152,10 @@ inline Microphone::~Microphone() {
     prepare_buffer_thread_.join();
 }
 
-inline void CALLBACK
-Microphone::callbackFunc(HWAVEIN hwavein, UINT uMsg, DWORD_PTR dwInstance,
-                         DWORD_PTR dwParam1, DWORD_PTR dwParam2) {
+inline void CALLBACK Microphone::callbackFunc(HWAVEIN hwavein, UINT uMsg,
+                                              DWORD_PTR dwInstance,
+                                              DWORD_PTR dwParam1,
+                                              DWORD_PTR dwParam2) {
   auto header = (LPWAVEHDR)dwParam1;
   auto mic = (header ? (Microphone *)header->dwUser : nullptr);
   if (!mic || !header)
